HW 1 part 1
This is the first homework that I've done. The hardest part was going through
docs and trying to find out what functions are available to me. I did know of
the existence of the compose operator, but I didn't know what it was. After I
found it in the docs I was able to make a the code more succinct. I think this
is what they talk about with functional programming. I made the program about
functions.

I also implemented the hw in Elixir. I don't know how I feel about it there, but
I did like that I had map_every.

Amos King @adkron <amos@binarynoggin.com>