HW 2 init
I have an issue with `buildMessage` not being exhaustive because it
needs to match on an empty map. I really didn't like building all this
stuff in the case statement and that is why I pulled it out. I think I
need to start thinking of functions returning functions and build it up
as arguments get passed in.

Amos King @adkron <amos@binarynoggin.com>